SimTop: Fix ISA table Bugs
