+++ src/main/scala/Core/IFU_IDU/ct_core.scala
@@ -261,22 +261,22 @@ class SimTop extends Module with Config with ROBConfig {
   io.uart.out.valid := false.B
   io.uart.out.ch    := 0.U
 
-  val instrCommit = Module(new DifftestInstrCommit)
-  instrCommit.io.clock := clock
-  instrCommit.io.coreid := 0.U
-  instrCommit.io.index := 0.U
-  instrCommit.io.skip := false.B
-  instrCommit.io.isRVC := false.B
-  instrCommit.io.scFailed := false.B
-
-  instrCommit.io.valid := true.B
-  instrCommit.io.pc    := 0.U
-
-  instrCommit.io.instr := 0.U
-
-  instrCommit.io.wen   := false.B
-  instrCommit.io.wdata := 0.U
-  instrCommit.io.wdest := 0.U
+  //  val instrCommit = Module(new DifftestInstrCommit)
+  //  instrCommit.io.clock := clock
+  //  instrCommit.io.coreid := 0.U
+  //  instrCommit.io.index := 0.U
+  //  instrCommit.io.skip := false.B
+  //  instrCommit.io.isRVC := false.B
+  //  instrCommit.io.scFailed := false.B
+  //
+  //  instrCommit.io.valid := true.B
+  //  instrCommit.io.pc    := 0.U
+  //
+  //  instrCommit.io.instr := 0.U
+  //
+  //  instrCommit.io.wen   := false.B
+  //  instrCommit.io.wdata := 0.U
+  //  instrCommit.io.wdest := 0.U
 
 
   val csrCommit = Module(new DifftestCSRState)
@@ -300,19 +300,5 @@ class SimTop extends Module with Config with ROBConfig {
   csrCommit.io.mideleg        := 0.U
   csrCommit.io.medeleg        := 0.U
 
-  val cycleCnt = RegInit(0.U(64.W))
-  cycleCnt := cycleCnt + 1.U
-  val instrCnt = RegInit(0.U(64.W))
-  when(instrCommit.io.valid){
-    instrCnt := instrCnt + 1.U
-  }
 
-  val trap = Module(new DifftestTrapEvent)
-  trap.io.clock    := clock
-  trap.io.coreid   := 0.U
-  trap.io.valid    := false.B
-  trap.io.code     := 0.U // GoodTrap
-  trap.io.pc       := 0.U
-  trap.io.cycleCnt := cycleCnt
-  trap.io.instrCnt := instrCnt
 }
